#4 - Add documentation to all package level functions and types
