+++ server/config/config.go
@@ -2,6 +2,7 @@ package config
 
 import "github.com/spf13/viper"
 
+//InitConfig sets the default config values and binds environment variables. Should be called at the start of the application.
 func InitConfig() {
 	viper.SetDefault("env", "local")
