+++ server/database/mongo_adapter/db_connection.go
@@ -15,8 +15,17 @@ import (
 	"go.mongodb.org/mongo-driver/mongo/readpref"
 )
 
+// OpenConnection opens the connection to MongoDB server using the fields from the database config.
+// Initializes the adapter's context and cancel function, as well as its client, database, and collection pointers.
+// Returns any errors.
 func (db *MongoAdapter) OpenConnection() error {
-	dbConfig := viper.GetStringMap("database")[viper.GetString("env")].(config.DatabaseConfig)
+	env := viper.GetString("env")
+	mapResult, ok := viper.GetStringMap("database")[env]
+	if !ok {
+		return errors.New("no database config found for environment " + env)
+	}
+
+	dbConfig := mapResult.(config.DatabaseConfig)
 
 	db.context, db.cancelFunc = context.WithCancel(context.Background())
 	db.timeout = dbConfig.Timeout
@@ -43,14 +52,15 @@ func (db *MongoAdapter) OpenConnection() error {
 	db.Client = client
 	db.Database = client.Database(dbName)
 	db.Migrations = db.Database.Collection("migrations")
-	db.Users = db.Database.Collection("users")
 
 	return nil
 }
 
+// CloseConnection closes the connection to the MongoDB server and resets its client, database, and connection pointers.
+// The adapter also calls its cancel function to cancel any child requests that may still be running.
+// None of the adapter's pointers or context should be used after calling this function.
+// Returns any errors.
 func (db *MongoAdapter) CloseConnection() error {
-	defer db.cancelFunc() //cancel any remaining requests that may still be running
-
 	ctx, cancel := db.CreateStandardTimeoutContext()
 	err := db.Client.Disconnect(ctx)
 	cancel()
@@ -59,9 +69,18 @@ func (db *MongoAdapter) CloseConnection() error {
 		return common.ChainError("error closing database connection", err)
 	}
 
+	db.cancelFunc() //cancel any remaining requests that may still be running
+
+	//clean up the resources
+	db.Client = nil
+	db.Database = nil
+	db.Migrations = nil
+
 	return nil
 }
 
+// Ping pings the MongoDB server to verify it can still be reached.
+// Returns an error if it cannot, or if any other errors are encountered.
 func (db *MongoAdapter) Ping() error {
 	ctx, cancel := db.CreateStandardTimeoutContext()
 	err := db.Client.Ping(ctx, readpref.Primary())
