+++ server/models/migration_model.go
@@ -6,11 +6,13 @@ import (
 	"github.com/google/uuid"
 )
 
+// Migration is the migration model.
 type Migration struct {
 	ID        uuid.UUID `bson:"id"`
 	Timestamp string    `bson:"timestamp"`
 }
 
+// CreateNewMigration creates a new migration with a new ID and the given timestamp.
 func CreateNewMigration(timestamp string) *Migration {
 	return &Migration{
 		ID:        uuid.New(),
@@ -18,6 +20,8 @@ func CreateNewMigration(timestamp string) *Migration {
 	}
 }
 
+// ValidateMigrationTimestamp validates the given timestamp is in a valid format.
+// Returns a ValidateError indicating its result.
 func ValidateMigrationTimestamp(timestamp string) ValidateError {
 	matched, _ := regexp.MatchString(`^\d{14}$`, timestamp)
 	if !matched {
@@ -27,6 +31,8 @@ func ValidateMigrationTimestamp(timestamp string) ValidateError {
 	return CreateModelValidValidateError()
 }
 
+// Validate validates the migration is a valid migration model.
+// Returns a ValidateError indicating its result.
 func (m Migration) Validate() ValidateError {
 	if m.ID == uuid.Nil {
 		return CreateValidateError(MigrationInvalidID, "id cannot be nil")
